LayoutTests/fast/viewport/viewport-37.html

<head>
    <title>Viewport META with "height=device-height, initial-scale=1".</title>
    <meta name="viewport" content="height=device-height, initial-scale=1">
    <script>
        function test() {
            if (window.layoutTestController) {
                layoutTestController.dumpAsText();
                layoutTestController.dumpConfigurationForViewport(320, 352);
            }
        }
    </script>
</head>
<body onload="test();">
