LayoutTests/fast/viewport/viewport-37-expected.txt

ALERT: viewport size 436.364x480 scale 1 with limits [0.733333, 5] and userScalable true
